API file cleanup
